ocr/requirements.txt
pdf2image
pytesseract
pillow
fitz